\name{alpha.Summary}
\alias{alpha.Summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Alpha Summary}
\description{
This function gives what Cronbach's alpha would be if the test item were removed
}
\usage{
alpha.Summary(columns)
}
\arguments{
  \item{columns}{Columns of a data frame of matrix containing response variables}
}

\author{Harold C. Doran}

\examples{
% xx <- simRasch(xx)
% alpha.Summary(xx$data)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.

\keyword{misc}
