\name{fit.Stats}
\alias{fit.Stats}
\title{Rasch Fit Statistics}
\description{Computes the Infit and Outfit Statistics
}
\usage{
fit.Stats(dat, params)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{Columns of dataframe with item responses}
  \item{params}{Rasch location parameters}
}
\details{This function supplements the jml function
}
\author{Harold C. Doran}
\examples{
## See the MP vignette for details
}
\keyword{misc}

