\name{pb}
\alias{pb}
\title{Point Biserial Correlations}
\description{Computes Point Biserial Correlations}
\usage{
pb(data, ...)
}
\arguments{
  \item{data}{Columns of dataframe containing only the item responses}
  \item{...} {Further arguments passed to cor}
}
\author{Harold C. Doran}
\examples{
## See MP vignette
}
\keyword{misc}

