\name{piPlot}
\alias{piPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Person Item Plot}
\description{
 Plot the density of items and persons together
}
\usage{
piPlot(dat, params)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{Data frame or matrix containing item responses}
  \item{params}{Item parameters}
}
\author{Harold C. Doran}
\examples{
% xx <- simRasch(200, 10)
% aa <- jml(xx$data)
% piPlot(xx$data, aa$params)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
