\name{alpha}
\alias{alpha}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Cronbach's Coefficient Alpha}
\description{
  Computes Cronbach's alpha. This reduces to KR-20 when the columns of the data matrix are dichotomous.
}
\usage{
alpha(columns)
}
\arguments{
  \item{columns}{Columns of a data frame or matrix.}
}
\details{
The columns of the data must only contain the response variables}
\author{Harold C. Doran}
\examples{
% alpha(data)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
