\name{classical}
\alias{classical}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Classical Item Analysis}
\description{
Gives a basic classical item analysis of the test items
}
\usage{
classical(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{A data frame or matrix with response variables}
}
\author{Harold C. Doran}
\examples{
% xx <- simRasch(200, 10)
% classical(xx$data)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
